requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.7
cycler==0.11.0
fonttools==4.28.2
idna==3.3
joblib==1.1.0
kiwisolver==1.3.2
lyricsgenius==3.0.1
matplotlib==3.5.0
numpy==1.21.4
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
plotly==5.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.3
seaborn==0.11.2
setuptools-scm==6.3.2
six==1.16.0
sklearn==0.0
soupsieve==2.3.1
spotipy==2.19.0
tenacity==8.0.1
threadpoolctl==3.0.0
tomli==1.2.2
urllib3==1.26.7